requirements.txt
aiogram~=3.6.0
asyncpg~=0.29.0
aiojobs~=1.2.1
tortoise-orm~=0.20.1
aerich~=0.7.2
redis~=5.1.0b5
hiredis~=2.3.2
loguru~=0.7.2
uvloop~=0.19.0
python-dateutil~=2.9.0
python-decouple~=3.8